feat(transport): treat empty tls config as disabled
## What
- treat an empty TLS config as disabled instead of configured
- keep server TLS setup opportunistic: load cert/key only when both are present and load CA when present
- remove the missing-key-pair sentinel and strict server constructor rejection tests

## Why
- configs that omit TLS material should continue to behave like TLS is not configured
- server config should not fail only because a decoded TLS block is empty

## Testing
- go test ./crypto/tls/config ./config/server ./transport/http ./transport/grpc ./config/client
- make lint

AI-assisted-by: [Codex](https://openai.com/codex/)
